영화추천사이트/movieback/requirements.txt
beautifulsoup4==4.6.0
certifi==2020.4.5.2
chardet==3.0.4
click==7.1.2
colorama==0.4.3
cycler==0.10.0
defusedxml==0.6.0
Django==2.1.15
django-allauth==0.42.0
django-cors-headers==3.3.0
django-extensions==2.2.9
django-rest-auth==0.9.5
djangorestframework==3.11.0
idna==2.9
joblib==0.15.1
JPype1==0.7.5
kiwisolver==1.2.0
konlpy==0.5.2
krwordrank==1.0.2
lxml==4.5.1
matplotlib==3.2.1
nltk==3.5
numpy==1.18.5
oauthlib==3.1.0
Pillow==7.1.2
pygame==1.9.6
pyparsing==2.4.7
pyperclip==1.8.0
PySocks==1.7.1
pytagcloud==0.3.5
python-dateutil==2.8.1
python3-openid==3.1.0
pytz==2020.1
regex==2020.6.8
requests==2.23.0
requests-oauthlib==1.3.0
scikit-learn==0.23.1
scipy==1.4.1
selenium==3.141.0
simplejson==3.17.0
six==1.15.0
sklearn==0.0
soupsieve==2.0.1
threadpoolctl==2.1.0
tqdm==4.46.1
tweepy==3.8.0
urllib3==1.25.9
wordcloud==1.7.0
